+++ testing/src/main/java/com/yubico/yubikit/testing/piv/PivDeviceTests.java
@@ -15,6 +15,10 @@
  */
 package com.yubico.yubikit.testing.piv;
 
+import static com.yubico.yubikit.testing.piv.PivTestConstants.DEFAULT_MANAGEMENT_KEY;
+import static com.yubico.yubikit.testing.piv.PivTestConstants.DEFAULT_PIN;
+import static com.yubico.yubikit.testing.piv.PivTestConstants.DEFAULT_PUK;
+
 import com.yubico.yubikit.core.Logger;
 import com.yubico.yubikit.core.application.BadResponseException;
 import com.yubico.yubikit.core.smartcard.ApduException;
@@ -28,10 +32,7 @@ import com.yubico.yubikit.piv.PivSession;
 import com.yubico.yubikit.piv.Slot;
 import com.yubico.yubikit.piv.TouchPolicy;
 import com.yubico.yubikit.piv.jca.PivAlgorithmParameterSpec;
-import com.yubico.yubikit.piv.jca.PivPrivateKey;
-import com.yubico.yubikit.piv.jca.PivProvider;
 
-import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.bouncycastle.util.encoders.Hex;
 import org.hamcrest.CoreMatchers;
 import org.hamcrest.MatcherAssert;
@@ -43,22 +44,9 @@ import java.security.InvalidAlgorithmParameterException;
 import java.security.InvalidKeyException;
 import java.security.KeyPair;
 import java.security.KeyPairGenerator;
-import java.security.KeyStore;
-import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
-import java.security.PrivateKey;
 import java.security.PublicKey;
-import java.security.Security;
-import java.security.Signature;
-import java.security.SignatureException;
-import java.security.UnrecoverableKeyException;
-import java.security.cert.CertificateException;
 import java.security.interfaces.ECPublicKey;
-import java.security.spec.AlgorithmParameterSpec;
-import java.security.spec.MGF1ParameterSpec;
-import java.security.spec.PSSParameterSpec;
-import java.util.Arrays;
-import java.util.List;
 
 import javax.crypto.BadPaddingException;
 import javax.crypto.Cipher;
@@ -67,11 +55,6 @@ import javax.crypto.KeyAgreement;
 import javax.crypto.NoSuchPaddingException;
 
 public class PivDeviceTests {
-    private static final byte[] DEFAULT_MANAGEMENT_KEY = Hex.decode("010203040506070801020304050607080102030405060708");
-    private static final char[] DEFAULT_PIN = "123456".toCharArray();
-    private static final char[] DEFAULT_PUK = "12345678".toCharArray();
-
-    private static final List<String> MESSAGE_DIGESTS = Arrays.asList("SHA-1", "SHA-224", "SHA-256", "SHA-384", "SHA-512");
 
     public static void testManagementKey(PivSession piv) throws BadResponseException, IOException, ApduException {
         byte[] key2 = Hex.decode("010203040102030401020304010203040102030401020304");
@@ -192,106 +175,6 @@ public class PivDeviceTests {
 
         // Change PUK
         piv.changePuk(puk2, DEFAULT_PUK);
-    }
-
-    public static void testSignAllHashes(PivSession piv, Slot slot, KeyType keyType, PublicKey publicKey) throws ApduException, NoSuchAlgorithmException, InvalidPinException, IOException, InvalidKeyException, BadResponseException {
-        for (String hash : MESSAGE_DIGESTS) {
-            testSign(piv, slot, keyType, publicKey, hash);
-        }
-    }
-
-    public static void testSign(PivSession piv, Slot slot, KeyType keyType, PublicKey publicKey, String digest) throws NoSuchAlgorithmException, IOException, ApduException, InvalidPinException, InvalidKeyException, BadResponseException {
-        byte[] message = "Hello world!".getBytes(StandardCharsets.UTF_8);
-
-        String signatureAlgorithm = digest.replace("-", "") + "with";
-        switch (keyType.params.algorithm) {
-            case RSA:
-                signatureAlgorithm += "RSA";
-                break;
-            case EC:
-                signatureAlgorithm += "ECDSA";
-                break;
-        }
-
-        Logger.d("Create signature");
-        //byte[] signature = piv.sign(slot, keyType, message, sig);
-        try {
-            Signature sig = Signature.getInstance(signatureAlgorithm);
-
-            KeyStore keyStore = KeyStore.getInstance("YKPiv");
-            keyStore.load(null);
-            PrivateKey privateKey = (PrivateKey) keyStore.getKey(Integer.toString(slot.value, 16), DEFAULT_PIN);
-            sig.initSign(privateKey);
-            sig.update(message);
-            byte[] signature = sig.sign();
-
-            // Verify
-            sig = Signature.getInstance(signatureAlgorithm);
-            sig.initVerify(publicKey);
-            sig.update(message);
-            Assert.assertTrue("Verify signature", sig.verify(signature));
-        } catch (InvalidKeyException | SignatureException | KeyStoreException | CertificateException | UnrecoverableKeyException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public static void testSign(PivSession piv, KeyType keyType) throws NoSuchAlgorithmException, IOException, ApduException, InvalidPinException, InvalidKeyException, BadResponseException, InvalidAlgorithmParameterException, SignatureException {
-        Security.addProvider(new BouncyCastleProvider());
-        Security.addProvider(new PivProvider(piv));
-        piv.authenticate(ManagementKeyType.TDES, DEFAULT_MANAGEMENT_KEY);
-        Logger.d("Generate key: " + keyType);
-
-        KeyPairGenerator kpg = KeyPairGenerator.getInstance(keyType.params.algorithm.name());
-        kpg.initialize(new PivAlgorithmParameterSpec(Slot.SIGNATURE, keyType, PinPolicy.DEFAULT, TouchPolicy.DEFAULT, DEFAULT_PIN));
-        KeyPair keyPair = kpg.generateKeyPair();
-
-        //PublicKey publicKey = piv.generateKey(Slot.SIGNATURE, keyType, PinPolicy.DEFAULT, TouchPolicy.DEFAULT);
-        //PrivateKey privateKey = PivPrivateKey.from(publicKey, Slot.SIGNATURE, DEFAULT_PIN);
-
-        switch (keyType.params.algorithm) {
-            case EC:
-                testSign(keyPair, "SHA1withECDSA", null);
-                testSign(keyPair, "SHA256withECDSA", null);
-                //noinspection SpellCheckingInspection
-                testSign(keyPair, "NONEwithECDSA", null);
-                testSign(keyPair, "SHA3-256withECDSA", null);
-                break;
-            case RSA:
-                testSign(keyPair, "SHA1withRSA", null);
-                testSign(keyPair, "SHA256withRSA", null);
-                testSign(keyPair, "RSASSA-PSS", new PSSParameterSpec("SHA-256", "MGF1", MGF1ParameterSpec.SHA256, 8, 1));
-
-                // Test with custom parameter. We use a 0-length salt and ensure signatures are the same
-                PSSParameterSpec param = new PSSParameterSpec("SHA-256", "MGF1", MGF1ParameterSpec.SHA256, 0, 1);
-                byte[] sig1 = testSign(keyPair, "RSASSA-PSS", param);
-                byte[] sig2 = testSign(keyPair, "RSASSA-PSS", param);
-                Assert.assertArrayEquals("PSS parameters not used, signatures are not identical!", sig1, sig2);
-                break;
-        }
-    }
-
-    public static byte[] testSign(KeyPair keyPair, String signatureAlgorithm, AlgorithmParameterSpec param) throws NoSuchAlgorithmException, IOException, ApduException, InvalidPinException, BadResponseException, InvalidAlgorithmParameterException, InvalidKeyException, SignatureException {
-        byte[] message = "Hello world!".getBytes(StandardCharsets.UTF_8);
-
-        Logger.d("Create signature using " + signatureAlgorithm);
-        Signature signer = Signature.getInstance(signatureAlgorithm);
-        signer.initSign(keyPair.getPrivate());
-        if (param != null) signer.setParameter(param);
-        signer.update(message);
-        byte[] signature = signer.sign();
-
-        //byte[] signature = piv.sign(Slot.SIGNATURE, KeyType.fromKey(publicKey), message, signatureAlgorithm);
-        try {
-            Signature verifier = Signature.getInstance(signatureAlgorithm);
-            verifier.initVerify(keyPair.getPublic());
-            if (param != null) verifier.setParameter(param);
-            verifier.update(message);
-            Assert.assertTrue("Verify signature", verifier.verify(signature));
-            Logger.d("Signature verified for: " + signatureAlgorithm);
-            return signature;
-        } catch (InvalidKeyException | SignatureException e) {
-            throw new RuntimeException(e);
-        }
     }
 
     public static void testDecrypt(PivSession piv, KeyType keyType) throws BadResponseException, IOException, ApduException, NoSuchPaddingException, NoSuchAlgorithmException, InvalidKeyException, BadPaddingException, IllegalBlockSizeException, InvalidPinException, InvalidAlgorithmParameterException {
